# master-data/src/rekening-gaji/entities/rekening-gaji.entity.ts
export class RekeningGaji {
  id: number;
  nip: string;
  nama_pegawai: string;
  nama_bank: string;
  nomor_rekening: string;
  cabang_bank: string;
  skpd_id: number;
  created_at: Date;
  updated_at: Date;
}